requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.0.1+cpu
pathlib==1.0.1
requests==2.31.0
uvicorn==0.21.1
fastapi==0.108.0
PyYAML==6.0
scikit-learn==1.3.0
cohere==4.35
numpy==1.26.4